also filter by baseline state for other check components
